% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-attr_names.R
\docType{data}
\name{attr_names}
\alias{attr_names}
\title{Attribute Name Metadata}
\format{
A list with names and values for the formal and in-code names of various attributes
}
\usage{
attr_names
}
\description{
This list contains the meta data for all of the attributes that are stratified accross for the EpiModel Simulations.
}
\keyword{list}
