% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_targ_from_dat.R
\name{build_targ_from_dat}
\alias{build_targ_from_dat}
\title{Build Targets from dat objec}
\usage{
build_targ_from_dat(dat)
}
\arguments{
\item{dat}{the dat object from an EpiModel Simulation}
}
\value{
A dataframe with all targets contained in the dat object matching the
structure of the WHAMP.targ dataframe.
}
\description{
Build Targets from dat objec
}
\examples{
# See vignette
}
