% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit_rates.R
\name{make_exit_plot}
\alias{make_exit_plot}
\title{Create a plot of the exiting rate of nodes in a population.}
\usage{
make_exit_plot(sim)
}
\arguments{
\item{sim}{The simulation for which you would like to plot the exit rates.}
}
\description{
This function calculates the mortality rates by comparing the population
distribution at different time steps.  For example if there are 50
individuals in the population age 30 at year 1990 and only 45 individuals
in the population age 31 at year 1991, we can determine that 5 individuals
in this age cohort died.
}
\details{
Note that because of the incredibly small mortality rates for
younger individuals, the plots for the younger individuals will be noisy.
}
