% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_counts.R
\name{make_num_plot}
\alias{make_num_plot}
\title{Make plot of overall population counts.}
\usage{
make_num_plot(sim)
}
\arguments{
\item{sim}{Simulation which to plot}
}
\description{
Make plot of overall population counts.
}
