% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_counts.R
\name{make_split_num_plot}
\alias{make_split_num_plot}
\title{Population counts}
\usage{
make_split_num_plot(sim, race_spec = FALSE, ofm_proj = NULL)
}
\arguments{
\item{sim}{The simulation you wish to see.}

\item{race_spec}{Are race specific populaiton counts desired.}

\item{ofm_proj}{If desired the OFM based population projections.}
}
\description{
Population counts split by active (Age < 65) and inactive (Age >= 65)
}
