% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epi.R
\name{plot_epi}
\alias{plot_epi}
\title{Create a plot of epi measures from EpiModel simulation}
\usage{
plot_epi(
  plot_data,
  brk_across = "ovr",
  targets = NULL,
  plot_type = "line",
  year_range = c(2000, 2030)
)
}
\arguments{
\item{plot_data}{Data object that is output from}

\item{brk_across}{the attribute over which you want to break out the measure
(for example race).}

\item{targets}{A dataframe that contains the targets for the measure of
interest.  Note that plot_data should be specific to a singe measure.}

\item{plot_type}{type of geom to call.  Default is for a line plot
(geom_line) when `plot_type == "line"`.  Otherwise `geom_smooth` is called.}

\item{year_range}{Range of years over which plot should be created.}
}
\value{
a ggplot object.
}
\description{
Create a plot of epi measures from EpiModel simulation
}
\examples{
# See vignette
}
