% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-WHAMP.targs.R
\docType{data}
\name{WHAMP.targs}
\alias{WHAMP.targs}
\title{Constructed and hard-coded targets}
\format{
A dataframe with targets for various epi values
}
\usage{
WHAMP.targs
}
\description{
This dataset contains many of the target parameters used when
constructing the diagnostic plots for the WHAMP EpiModel Simulation.
}
\keyword{dataframe}
